+++ doctr/models/modules/__init__.py
@@ -1,1 +1,2 @@
 from .transformer import *
+from vision_transformer import *
