+++ doctr/models/modules/transformer/pytorch.py
@@ -11,33 +11,7 @@ from typing import Optional, Tuple
 import torch
 from torch import nn
 
-__all__ = ["Decoder", "PositionalEncoding", "PatchEmbedding"]
-
-
-class PatchEmbedding(nn.Module):
-    """Compute 2D patch embedding"""
-
-    def __init__(
-        self,
-        img_size: Tuple[int],
-        patch_size: Tuple[int] = (4, 8),  # different from paper to match with 32x128 input
-        channels: int = 3,
-        embed_dim: int = 768,
-    ) -> None:
-
-        super().__init__()
-        self.img_size = img_size
-        self.patch_size = (img_size[0] // patch_size[0], img_size[1] // patch_size[1])
-        self.proj = nn.Conv2d(channels, embed_dim, kernel_size=patch_size, stride=patch_size)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        B, C, H, W = x.shape
-
-        assert H % self.patch_size[0] == 0, "Image height must be divisible by patch height"
-        assert W % self.patch_size[1] == 0, "Image width must be divisible by patch width"
-
-        x = self.proj(x)
-        return x.flatten(2).transpose(1, 2)  # BCHW -> BNC
+__all__ = ["Decoder", "PositionalEncoding"]
 
 
 class PositionalEncoding(nn.Module):
